feat(TaskList): display tasks from database
-creates a TaskItem composable for the list

+++ app/src/main/java/com/paulchibamba/deeply/navigation/Screens.kt
@@ -12,7 +12,7 @@ class Screens(navController: NavHostController) {
         }
     }
 
-    val taskDetail: (Int) -> Unit = { taskId ->
+    val taskDetail: (Long) -> Unit = { taskId ->
         navController.navigate("task_detail/$taskId")
     }
 }
